refactor: improve code ss-011
