+++ src/libs/constants/constants.ts
@@ -1,2 +1,2 @@
 export { EMPTY_ARRAY_LENGTH } from "./empty-array-length.js";
-export { UI_INDEX_BASE } from "./ua-index-base.js";
+export { UI_INDEX_BASE } from "./ui-index-base.js";
